% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_rcar.R
\name{cv.rcar}
\alias{cv.rcar}
\title{Cross-validation for rcar}
\usage{
cv.rcar(data, y.rank = 1, s = 0.3, c = 0.7, maxl = 6)
}
\arguments{
\item{data:}{A categorical data set with binary response variable; each row is an observation vector.}

\item{y.rank:}{The rank of the response variable.}

\item{s:}{A user supplied value for the minimal support of an item set.}

\item{c:}{A user supplied value for the minimal confidence of class-association rules.}

\item{maxl:}{A user specified value for the maximal number of items per item set (default: 6 items).}
}
\value{
An object of class "cv.rcar" is returned, which is a list with the ingredients of the cross-validation fit plus the elements of the most accurate model.
}
\description{
Does 10-fold cross-validation for RCAR over an automatically choosen sequence of values of lambda.
}
\details{
10-folds cross validation is calculated over an automatically determined sequence of lambda values ranging from the
value of lambda such that all the coefficients are zero down to value of lambda whose the deviance do not change
from lambda to the next.
Note also that the results of cv.rcar are random, since the folds are selected randomly.
Users can reduce this randomness by running cv.rcar many times, and averaging the error curves.
}
